+++ CHANGELOG.md
@@ -3,7 +3,6 @@
 ## TODO LIST
 
 - 建造者设计模式的熟悉（常见的设计模式总结）
-- SignAuthInterceptor  使用SDK  的sample和例子，简化对接成本
 - Elasticsearch 深入研究，面试，实战
 - [代码在线生成器](https://gitee.com/zhangquansheng/magic/tree/4.2.0-springboot-code-gen/)
 - SOLID 架构设计 深入学习
@@ -17,8 +16,14 @@
 - **需要花费很长时间的事情，需要慎重考虑一下在去做**
 - **对未来乐观**
 
-
-
+## [v4.3.0] 2020.5.7
+- `SignAuthInterceptor` 简化对接成本;
+    - [SignAuthFeignConfig](https://gitee.com/zhangquansheng/magic/blob/4.2.0-springboot-code-gen/src/main/java/com/zhengcheng/magic/common/config/SignAuthFeignConfig.java)
+- 增加 `SignAuthUtils` 工具类
+- `zc-feign-spring-boot-starter` 远程通信通用组件，默认配置中增加SignAuth的参数，详细见`com.zhengcheng.feign.FeignAutoConfiguration`,其中默认的秘钥为：security.api.key，请在配置文件中写入秘钥，默认为 `zhengcheng` 
+- 技术分享 [API接口防止参数篡改和重放攻击](https://note.youdao.com/ynoteshare1/index.html?id=ed15f29e7ad1ff2d15a1236231283bc7&type=note)
+- 删除 `FeignInterceptorConfig` , 默认配置中增加 Feign OAuth2 拦截器
+- `TraceIdInterceptor` 增加 `applicationName` 构造参数，修复无法打印项目名的问题
 
 ## [v4.4.0] 2020.4.18